Splits invoervelden.class.php
