+++ lib/view/formulier/invoervelden/LandField.class.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * LandField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * LandField met een aantal autocomplete suggesties voor landen.
+ * Doet verder geen controle op niet-bestaande landen...
+ */
+class LandField extends TextField {
+
+	public function __construct($name, $value, $description) {
+		parent::__construct($name, $value, $description);
+		$this->suggestions[] = array('Nederland', 'België', 'Duitsland', 'Frankrijk', 'Verenigd Koninkrijk', 'Verenigde Staten');
+	}
+
+}
