+++ lib/view/formulier/invoervelden/RequiredDuckField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredDuckField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see DuckField
+ */
+class RequiredDuckField extends DuckField {
+
+	public $required = true;
+
+}
