+++ lib/view/formulier/invoervelden/RequiredEmailField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredEmailField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see EmailField
+ */
+class RequiredEmailField extends EmailField {
+
+	public $required = true;
+
+}
