+++ lib/view/formulier/invoervelden/RequiredEntityField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredEntityField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see EntityField
+ */
+class RequiredEntityField extends EntityField {
+
+	public $required = true;
+
+}
