+++ lib/view/formulier/invoervelden/RequiredFileNameField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredFileNameField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see FileNameField
+ */
+class RequiredFileNameField extends FileNameField {
+
+	public $required = true;
+
+}
