+++ lib/view/formulier/invoervelden/RequiredIBANField.class.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * RequiredIBANField.class.php
+ *
+ * @author Jorai Rijsdijk <jorairijsdijk@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see IBANField
+ */
+class RequiredIBANField extends IBANField {
+	public $required = true;
+}
