+++ lib/view/formulier/invoervelden/RequiredLandField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredLandField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see LandField
+ */
+class RequiredLandField extends LandField {
+
+	public $required = true;
+
+}
