+++ lib/view/formulier/invoervelden/RequiredLidField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredLidField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see LidField
+ */
+class RequiredLidField extends LidField {
+
+	public $required = true;
+
+}
