+++ lib/view/formulier/invoervelden/RequiredRechtenField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredRechtenField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see RechtenField
+ */
+class RequiredRechtenField extends RechtenField {
+
+	public $required = true;
+
+}
