+++ lib/view/formulier/invoervelden/RequiredStudieField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredStudieField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see StudieField
+ */
+class RequiredStudieField extends StudieField {
+
+	public $required = true;
+
+}
