+++ lib/view/formulier/invoervelden/RequiredTextField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredTextField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see TextField
+ */
+class RequiredTextField extends TextField {
+
+	public $required = true;
+
+}
