+++ lib/view/formulier/invoervelden/RequiredTextareaField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredTextareaField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see TextareaField
+ */
+class RequiredTextareaField extends TextareaField {
+
+	public $required = true;
+
+}
