+++ lib/view/formulier/invoervelden/RequiredUrlField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredUrlField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see UrlField
+ */
+class RequiredUrlField extends UrlField {
+
+	public $required = true;
+
+}
