+++ lib/view/formulier/invoervelden/RequiredUsernameField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredUsernameField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see UsernameField
+ */
+class RequiredUsernameField extends UsernameField {
+
+	public $required = true;
+
+}
