+++ lib/view/formulier/invoervelden/RequiredWachtwoordField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredWachtwoordField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see WachtwoordField
+ */
+class RequiredWachtwoordField extends WachtwoordField {
+
+	public $required = true;
+
+}
