+++ lib/view/formulier/invoervelden/RequiredWachtwoordWijzigenField.class.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * RequiredWachtwoordWijzigenField.class.php
+ *
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see WachtwoordWijzigenField
+ */
+class RequiredWachtwoordWijzigenField extends WachtwoordWijzigenField {
+
+	public $required = true;
+
+}
